[BUGFIX] Wrong column configuration for XML-type data
The configuration validation for columns of XML-type data should
not report an error when all node-mapping properties are missing
since the import process is fine with this and takes the value
of the node itself.

Report a notice instead, just to make sure this is what the user
really wanted.

Resolves: #5