[FEATURE] Add functional testing
Functional testing is introduced for the Importer class, based on
data from the "externalimport_test" extension, resolves #7.

An API is added to force the setting of the storage pid, necessary
during running the tests.

NOTE: one test has revealed a bug and is still failing at this point.
A related issue has been opened.